backend-users/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-jose
python-keycloak
requests
redis